Fix/org creation (#17)
* feat(org): prevent duplicate org creation with 409 conflict

* feat(clerk): update setOrg service to use string orgId instead of uint

This is done to prevent converting string to uint type in setOrg

* fix(org): pass user from context to createOrg usecase in create handler

Allows the usecase to assign the org to the user after creation and set orgId on the Clerk user.

* feat(org): add transaction support for creating and assigning org to user

* feat(org): add internalMsg to api error in org creation service

* fix(user): remove extra where clause in update, rely on gorm Updates with model ID

Previously, the repository update used two IDs in the query by combining
a where clause with gorm's Updates method. This was redundant and could
lead to unexpected behavior.

Now, the update relies solely on gorm Updates, which automatically uses
the ID present in the user model.

* fix(routes): add clerk service and app context to usecase constructor

* fear(error): add error messages

* feat(identity): set clerk ID in custom claims and context user

Adds support for propagating the clerk ID by including it in both
JWT custom claims and the context user object. This allows services
to reliably access the clerk ID during request handling.

* refactor(types): update `CreateOrgInput` to take context user instead of `UserID`

* fix(test): fix identity test