Feat/update user and webhook module (#36)
* feat(user): update user service to return error instead of apiError type

Update handlers and repository to reflect change in service

This is done to simplify service and make it reusable

* feat(clerk): move clerk into a package outside module to make mocking service easier

update import in files

* feat(webhook): update service to return error instead of apiError type

Update handler to reflect change in service

This is done to simplify service and make it reusable

* feat(middleware): relocate verify-webhook middleware to middleware package

Move the verify-webhook middleware into the dedicated middleware package to
make testing easier. Updated all dependent files to
use the new location.

* feat(interfaces): remove clerk interface

- updateuser and webhook interface

* feat(test): add webhook integration test

* feat(swagger): update docs